// src/interfaces/constant/map.interface.ts
//=====================================
//
//  ###    ###    ###    #####    ####
//  ## #  # ##   ## ##   ##  ##  ##
//  ##  ##  ##  ##   ##  #####    ###
//  ##      ##  #######  ##         ##
//  ##      ##  ##   ##  ##      ####
//
//=====================================

export interface IMapCustomerType {
  erpName: string;
  name: string;
  color: string;
}